Fix palette color 255 showing as black
Was due to it being 1.0 in floating point representation which indexed off the end of the palette texture resulting in black. I also noticed that it might have been the case that some other colors weren't snapping to the correct palette index. E.g. seemed 254 was showing 253 and 253 showing 252.

Use usampler2D and texelFetch for exact mapping of the screen's R8 byte values to palette indices.